# api/openapi.yaml
openapi: 3.1.0
info:
  title: Island Media Co connected assessment API
  version: 2.0.0
  description: >-
    Fictional mock boundary for the Frameworks technical assessment. Protected
    prototype endpoints use X-Prototype-User-Id to represent the selected
    seeded or newly registered user. Production authentication is outside scope.
servers:
  - url: /api
paths:
...
  /session/switch:
    post:
      summary: Switch to a seeded prototype user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
      responses:
        "200":
          description: Prototype session selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
  /products:
    get:
      summary: Search the public product catalogue for requested dates
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: mediaType
          schema:
            type: string
        - in: query
          name: locationId
          schema:
            type: string
        - in: query
          name: maxMonthlyBudget
          schema:
            type: number
            minimum: 0
      responses:
        "200":
          description: Matching products with calculated availability
          content:
            application/json:
              schema:
                type: object
                required: [items, query]
                properties:
                  query:
                    $ref: "#/components/schemas/ProductSearchQuery"
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductSearchResult"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /products/{productId}:
    get:
      summary: Get one product with availability detail
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /mobile/work-orders:
    get:
      summary: List work orders assigned to the selected fitter
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Assigned mobile work orders
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkOrder"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    PrototypeUserId:
      in: header
      name: X-Prototype-User-Id
      required: true
      schema:
        type: string
...
  schemas:
...
    Session:
      type: object
      required: [user, organisation]
      properties:
        user:
          $ref: "#/components/schemas/User"
        organisation:
          oneOf:
            - $ref: "#/components/schemas/Organisation"
            - type: "null"
    User:
      type: object
      required: [id, name, email, role, status]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [client, manager, fitter]
        organisationId:
          type: [string, "null"]
        status:
          type: string
    Organisation:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        contractCount:
          type: integer
          minimum: 0
    ProductSearchQuery:
      type: object
      required: [startDate, endDate]
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        mediaType:
          type: [string, "null"]
        locationId:
          type: [string, "null"]
        maxMonthlyBudget:
          type: [number, "null"]
    IndicativeRate:
      type: object
      required: [currency, label]
      properties:
        currency:
          type: string
          const: GBP
        amount:
          type: [number, "null"]
        unit:
          type: [string, "null"]
        monthlyEquivalent:
          type: [number, "null"]
        label:
          type: string
    AvailabilitySummary:
      type: object
      required: [state, reason, calculatedAt]
      properties:
        state:
          type: string
          enum: [available, unavailable, confirmation_required]
        reason:
          type: string
        calculatedAt:
          type: string
          format: date-time
        availableAssetCount:
          type: [integer, "null"]
        availableCapacity:
          type: [integer, "null"]
        totalCapacity:
          type: [integer, "null"]
        freshestVerificationAt:
          type: [string, "null"]
          format: date-time
    ProductSearchResult:
      type: object
      required:
        [
          id,
          name,
          mediaOwnerName,
          mediaType,
          locationNames,
          allocationModel,
          indicativeRate,
          minimumTermDays,
          availability,
        ]
      properties:
        id:
          type: string
        name:
          type: string
        mediaOwnerName:
          type: string
        mediaType:
          type: string
        locationNames:
          type: array
          items:
            type: string
        allocationModel:
          type: string
          enum: [exclusive_asset, capacity_pool]
        indicativeRate:
          $ref: "#/components/schemas/IndicativeRate"
        minimumTermDays:
          type: integer
        availability:
          $ref: "#/components/schemas/AvailabilitySummary"
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/ProductSearchResult"
        - type: object
          required: [description, creativeSpec, assetOptions]
          properties:
            description:
              type: string
            creativeSpec:
              type: [object, "null"]
              additionalProperties: true
            assetOptions:
              type: array
              items:
                $ref: "#/components/schemas/AssetOption"
    AssetOption:
      type: object
      required: [id, name, status, availability]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        verifiedAt:
          type: [string, "null"]
          format: date-time
        verificationSource:
          type: [string, "null"]
        note:
          type: [string, "null"]
        availability:
          $ref: "#/components/schemas/AvailabilitySummary"
...
    HistoryEntry:
      type: object
      required: [at, actor, action]
      properties:
        at:
          type: string
          format: date-time
        actor:
          type: string
        action:
          type: string
        note:
          type: [string, "null"]
...
    WorkOrder:
      type: object
      required:
        [
          id,
          campaignId,
          contractId,
          organisationId,
          type,
          status,
          assignedUserId,
          assetId,
          scheduledStart,
          scheduledEnd,
          locationLabel,
          instructions,
          history,
          proofRecordIds,
        ]
      properties:
        id:
          type: string
        campaignId:
          type: string
        contractId:
          type: string
        organisationId:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [draft, assigned, travelling, on_site, blocked, completed]
        assignedUserId:
          type: string
        assetId:
          type: string
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        locationLabel:
          type: string
        instructions:
          type: string
        completionNote:
          type: [string, "null"]
        proofRecordIds:
          type: array
          items:
            type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: [object, "null"]
          additionalProperties: true
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Record not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The selected prototype user cannot access this record or action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: FORBIDDEN
            message: You cannot access this organisation's record.
...
    ServiceUnavailable:
      description: Simulated or real temporary service failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: SERVICE_UNAVAILABLE
            message: Please try again. The service is temporarily unavailable.